requirements.txt
pyTelegramBotAPI==4.4.0
Pillow==9.0.1